+++ login_stateful/lib/src/screens/login_screen.dart
@@ -11,8 +11,9 @@ class LoginScreenState extends State<LoginScreen> {
       child: Column(
         children: [
           emailField(),
-          // passwordField(),
-          // submitButton()
+          passwordField(),
+          Container(margin: EdgeInsets.only(top: 25.0),),
+          submitButton()
         ],
       ),
     );
@@ -20,6 +21,7 @@ class LoginScreenState extends State<LoginScreen> {
 
   Widget emailField() {
     return TextFormField(
+      keyboardType: TextInputType.emailAddress,
       decoration: InputDecoration(
         labelText: 'Email Adress',
         hintText: 'you@example.com'
@@ -27,7 +29,20 @@ class LoginScreenState extends State<LoginScreen> {
     );
   }
 
-  Widget passwordField() {}
+  Widget passwordField() {
+    return TextFormField(
+      decoration: InputDecoration(
+        labelText: 'Insert Password',
+        hintText: 'Password'
+      ),
+    );
+  }
 
-  Widget submitButton() {}
+  Widget submitButton() {
+    return RaisedButton(
+      color: Colors.blue,
+      child: Text('Submit'),
+      onPressed: () {},
+    );
+  }
 }